Propagate EOF found while checking trailing whitespace
Currently, if the source iterator returns None while the parser is
checking for whitespace before EOF, that None is ignored and the
iterator is expected to continue returning None after the whitespace
check, i.e., the iterator is expected to be fused.

In general, the iterator may not be fused, and calling next() after the
terminating None may lead to, for example, an unnecessary read from
blocking IO that hangs indefinitely while waiting for more data.

This change makes the EOF check aware if the iterator returns None in
the trailing whitespace check. Since all uses of eof() are preceded by a
call to parse_whitespace(), that call is moved into eof() and checked
there.