deserialize_bytes(): fix bug: clear scratch before parsing next
This bug can be observed if you parse a list of ByteBufs and two
elements have escape sequences so that the scratch space has to be used.

Note that the bug does not happen if you use JSON objects, since they keys
will be parsed as strings which triggers clearing the scratch space.