Remove irrelevant detail from Map::insert documentation
This sentence is from the documentation of HashMap::insert, which
applies to arbitrary key types K. In serde_json::Map the key type is
always String. Technically the behavior may still be observable if
someone looks at the memory address of the key string bytes through
`as_str() as *const str` but users are unlikely to require this
guarantee.