Change mapping of RawValue to a struct with one key
This is what serde_json::Number does in arbitrary precision mode. It is
more robust because it fails fast in formats that are not JSON. In the
previous implementation, it was too easy when mixing different data
formats to end up with a RawValue holding something that is not valid
JSON data:

    extern crate serde;
    extern crate serde_json;

    use serde::de::{Deserialize, IntoDeserializer, value::Error};
    use serde_json::RawValue;

    fn main() -> Result<(), Error> {
        let bad = RawValue::deserialize("~~~".to_owned().into_deserializer())?;
        println!("{}", bad);
        Ok(())
    }

The new implementation detects in this case that we are not
deserializing from a deserializer that speaks the language of
serde_json::RawValue and will fail fast rather than producing an illegal
RawValue.