Add an ArrayDeserializer to read a JSON array as a stream
This mimics the StreamDeserializer API and implements issue #404. Unlike
the StreamDeserializer, the ArrayDeserializer struct itself does not
keep track of the type of the array's elements, instead the next()
itself is generic to support deserialization of arrays with values of
different types. Unfortunately, this means we can't implement the
Iterator trait.