Require serde >= 1.0.100 due to no_std-related Error re-export
Relevant Serde PR: https://github.com/serde-rs/serde/pull/1620

To support both no-/std builds without using somewhat noisy
conditional compilation directives, we implement the re-exported
`serde::de::StdError` trait in https://github.com/serde-rs/json/pull/606.

However, this was only introduced in >= 1.0.100, so we need to bump
the version requirement of serde.

On the off chance of someone pulling in incompatible 1.0.4{5,6} versions
of serde_json, I believe it'd be good to yank those and cut a new
release with this patch.

Sorry for the omission in the original PR.

Fixes #612.