Fix target arch detection for limb width
cfg(target_arch) refers to the target of the current code being
compiled. In the case of the build script, that's the host architecture.
Cargo provides a separate env variable to build scripts to determine the
target arch of the library as opposed to the target arch of the build
script.