Allow serializing to fmt::Write (part 1: meat)
Since ser::Serializer writes to an implementer of io::Write,
the fact that the Serializer is producing valid utf-8
is lost to the type system. This prevents users from using
instances of fmt::Write with ser::Serializer.

This commit adds a ser::Write trait which is automatically
implemented by io::Write instances, and which can be implemented
by fmt::Write instances using the FmtWrite wrapper.

A large number of mechanical changes are required in order
to get ser::Serializer and ser::Formatter to use ser::Write
rather than io::Write. For ease of review, this commit only
contains the new traits or functions that needed refactoring.
The next commit contains the mechanical changes.